Move server-side error interfaces to the `server` module
