+++ server/src/main/java/io/spine/server/aggregate/ImportValidator.java
@@ -21,7 +21,7 @@
 package io.spine.server.aggregate;
 
 import io.spine.base.EventMessage;
-import io.spine.core.MessageInvalid;
+import io.spine.server.MessageInvalid;
 import io.spine.server.bus.EnvelopeValidator;
 import io.spine.server.type.EventEnvelope;
 import io.spine.validate.ConstraintViolation;
