+++ server/src/main/java/io/spine/server/bus/ValidatingFilter.java
@@ -25,7 +25,7 @@ import com.google.protobuf.Message;
 import io.spine.base.Error;
 import io.spine.base.Identifier;
 import io.spine.core.Ack;
-import io.spine.core.MessageInvalid;
+import io.spine.server.MessageInvalid;
 import io.spine.server.type.MessageEnvelope;
 
 import java.util.Optional;
