+++ server/src/main/java/io/spine/server/commandbus/CommandValidator.java
@@ -26,8 +26,8 @@ import io.spine.annotation.Internal;
 import io.spine.base.CommandMessage;
 import io.spine.core.Command;
 import io.spine.core.CommandId;
-import io.spine.core.MessageInvalid;
 import io.spine.core.TenantId;
+import io.spine.server.MessageInvalid;
 import io.spine.server.bus.EnvelopeValidator;
 import io.spine.server.type.CommandEnvelope;
 import io.spine.validate.ConstraintViolation;
